requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web automation
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.0

# Task scheduling
schedule>=1.2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0

# AI/LLM
openai>=1.0.0

# Database
sqlalchemy>=2.0.0

# Web framework (optional, for dashboard)
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
pyperclip>=1.8.2
